+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow dumping unsolved SMT files via `--dump-unsolved`
 - Allow resolving unknown addresses to only those that are already deployed
   via option `--only-deployed`
+- New empty solver option that simply makes sure that all SMT queries return
+  unknown
 
 ## Fixed
 - We now extract more Keccak computations than before from the Props to assert